fix: clear stale callbacks
